modal fix
